wrem/requirements/base.txt
coreapi==2.3.0
Django==1.11.4
django-admin-honeypot==1.0.0
django-axes==2.3.3
django-crispy-forms==1.6.1
django-filter==1.0.4
django-guardian==1.4.9
django-ordered-model==1.4.1
django-polymorphic==1.3
django-simple-email-confirmation==0.22
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drf-flex-fields==0.2.2
Markdown==2.6.8
Pillow==3.4.2
psycopg2==2.6.2
sqlparse==0.2.3
